requirements.txt
# Default Python stack for WeatherWear:
# FastAPI API + LangChain/LangGraph workflow
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.12.5
typing_extensions==4.15.0
requests==2.32.5
python-dotenv==1.2.2

langchain>=0.3
langchain-core>=0.3
langchain-openai>=0.2
langgraph>=0.2
chromadb>=0.5,<0.7
